+++ SofaKernel/modules/SofaHelper/src/sofa/helper/io/XspLoader.cpp
@@ -23,6 +23,7 @@
 #include <sofa/helper/system/FileRepository.h>
 #include <sofa/helper/system/Locale.h>
 #include <sofa/defaulttype/Vec.h>
+#include <sofa/helper/rmath.h>
 using sofa::defaulttype::Vector3;
 
 #include <sofa/core/objectmodel/Base.h>
